requirements.txt
requests
python-dotenv
discord
asyncio
pytz
bs4
